agno_basics/requirements.txt
agno
openai
pydantic
python-dotenv